Add minimum number of players to the IRoomInformation and make players byte

+++ Core/Project Eternity Online/Commands/Game Commands/Ask Join Room Script Server.cs
@@ -44,7 +44,7 @@ namespace ProjectEternity.Core.Online
                 else if (Owner.DicLocalRoom.TryGetValue(RoomID, out LocalRoom) && LocalRoom.Room.ListOnlinePlayer.Count < LocalRoom.Room.MaxNumberOfPlayer)
                 {
                     Owner.DicLocalRoom[RoomID].Room.AddOnlinePlayer(Sender, "Player");
-                    Owner.Database.UpdatePlayerCountInRoom(RoomID, LocalRoom.Room.ListOnlinePlayer.Count);
+                    Owner.Database.UpdatePlayerCountInRoom(RoomID, (byte)LocalRoom.Room.ListOnlinePlayer.Count);
                     Owner.ListPlayerToRemove.Add(Sender);
 
                     OnJoinRoomLocal(Sender, RoomID, Owner.DicLocalRoom[RoomID]);
